Make it possible to just compile for x86_64 simulator.

+++ platforms/ios/build.sh
@@ -13,8 +13,7 @@ fi
 
 cd ../..
 
-if [ "$1" != "simulator" ]; then    
-
+if [ \( "$1" != "simulator" \) -a \( "$1" != "simulator_x86_64" \) ]; then    
     mkdir -p platforms/ios/armv7s     
     cd platforms/ios/armv7s
     ../configure-device_xcode.sh -DIOS_ARCH="armv7s" $*
@@ -43,25 +42,28 @@ if [ "$1" != "simulator" ]; then
     cd ../../..
 fi
 
-
-mkdir -p platforms/ios/i386
-cd platforms/ios/i386
-if [ "$1" == "simulator" ]; then    
-  ../configure-simulator_xcode.sh -DIOS_ARCH="i386" -DBUILD_IPP_IW:BOOL=FALSE -DWITH_IPP:BOOL=FALSE ${@:2}
-else
-  ../configure-simulator_xcode.sh -DIOS_ARCH="i386" -DBUILD_IPP_IW:BOOL=FALSE -DWITH_IPP:BOOL=FALSE $*
+if [ "$1" != "simulator_x86_64" ]; then
+    mkdir -p platforms/ios/i386
+    cd platforms/ios/i386
+    if [ "$1" == "simulator" ]; then    
+  #skip the first parameter
+	../configure-simulator_xcode.sh -DIOS_ARCH="i386" -DBUILD_IPP_IW:BOOL=FALSE -DWITH_IPP:BOOL=FALSE ${@:2}
+    else
+	../configure-simulator_xcode.sh -DIOS_ARCH="i386" -DBUILD_IPP_IW:BOOL=FALSE -DWITH_IPP:BOOL=FALSE $*
+    fi
+    xcodebuild IPHONEOS_DEPLOYMENT_TARGET=8.0 -parallelizeTargets -jobs 8 -sdk iphonesimulator -configuration Release ARCHS="i386" -target ALL_BUILD clean build
+    cp -r ../../../libs/Release/* bin/Release
+    cp -r opencv/3rdparty/lib/Release/* bin/Release  
+    #cp -r opencv/3rdparty/ippicv/ippiw_mac/lib/ia32/* bin/Release
+    #cp -r opencv/3rdparty/ippicv/ippicv_mac/lib/ia32/* bin/Release
+    libtool -static -o libemgucv_i386.a bin/Release/*.a
+    cd ../../..
 fi
-xcodebuild IPHONEOS_DEPLOYMENT_TARGET=8.0 -parallelizeTargets -jobs 8 -sdk iphonesimulator -configuration Release ARCHS="i386" -target ALL_BUILD clean build
-cp -r ../../../libs/Release/* bin/Release
-cp -r opencv/3rdparty/lib/Release/* bin/Release  
-#cp -r opencv/3rdparty/ippicv/ippiw_mac/lib/ia32/* bin/Release
-#cp -r opencv/3rdparty/ippicv/ippicv_mac/lib/ia32/* bin/Release
-libtool -static -o libemgucv_i386.a bin/Release/*.a
-cd ../../..
 
 mkdir -p platforms/ios/x86_64
 cd platforms/ios/x86_64
-if [ "$1" == "simulator" ]; then    
+if [ \( "$1" == "simulator" \) -o \( "$1" == "simulator_x86_64" \) ]; then
+  #skip the first parameter    
   ../configure-simulator_xcode.sh -DIOS_ARCH="x86_64" -DBUILD_IPP_IW:BOOL=FALSE -DWITH_IPP:BOOL=FALSE ${@:2}
 else
   ../configure-simulator_xcode.sh -DIOS_ARCH="x86_64" -DBUILD_IPP_IW:BOOL=FALSE -DWITH_IPP:BOOL=FALSE $*
@@ -78,7 +80,10 @@ cd ../../..
 rm -rf platforms/ios/universal
 mkdir -p platforms/ios/universal
 if [ "$1" == "simulator" ]; then
+    #skip the first parameter
     lipo -create -output platforms/ios/universal/libemgucv.a platforms/ios/i386/libemgucv_i386.a platforms/ios/x86_64/libemgucv_x86_64.a
+elif [ "$1" == "simulator_x86_64" ]; then
+    cp -f platforms/ios/x86_64/libemgucv_x86_64.a platforms/ios/universal/libemgucv.a 
 else
     lipo -create -output platforms/ios/universal/libemgucv.a platforms/ios/armv7/libemgucv_armv7.a platforms/ios/armv7s/libemgucv_armv7s.a platforms/ios/arm64/libemgucv_arm64.a platforms/ios/i386/libemgucv_i386.a  platforms/ios/x86_64/libemgucv_x86_64.a
 fi
